# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: TUBA Product API
  version: 0.1.0
  description: Current API contract. Operations marked planned are design contracts and are not implemented endpoints.
servers:
  - url: /api/v1
security:
  - oidc: []
paths:
...
  /sources:
    get:
      operationId: listSources
      x-required-permission: source:manage
      summary: List source instances in the authenticated tenant
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: Source configuration without credentials
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/SourceInstance"}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: registerSource
      x-required-permission: source:manage
      summary: Register a source and issue its API key once
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SourceRegistrationInput"}
      responses:
        "201":
          description: Registered source; api_key is only returned at creation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RegisteredSource"}
        "400": {description: Invalid registration or release is not active/staged}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Source registry unavailable}
  /sources/{id}:
    delete:
      operationId: revokeSource
      x-required-permission: source:manage
      summary: Disable a source credential
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: "^src_[a-f0-9]{32}$"}}
      responses:
        "204": {description: Source disabled}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Source not found}
        "503": {description: Source registry unavailable}
  /sources/{id}/rotate:
    post:
      operationId: rotateSourceCredential
      x-required-permission: source:manage
      summary: Rotate a source key without changing the source epoch
      description: The previous key remains valid for 24 hours for queue draining. Store the returned API key; it is shown only once.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: "^src_[a-f0-9]{32}$"}}
      responses:
        "200":
          description: Updated source and newly issued one-time API key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RegisteredSource"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Source not found or disabled}
        "503": {description: Source registry unavailable}
  /collectors:
    get:
      operationId: listCollectors
      x-required-permission: source:manage
      summary: List registered Collectors in the authenticated tenant
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: Collector registry and last reported status
          content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: "#/components/schemas/CollectorSummary"}}}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Collector registry unavailable}
  /collectors/enrollments:
    post:
      operationId: createCollectorEnrollment
      x-required-permission: source:manage
      summary: Create a one-time enrollment token for a Collector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expires_in_minutes: {type: integer, minimum: 5, maximum: 1440, default: 30}
      responses:
        "201":
          description: Token returned once; response is no-store
          content: {application/json: {schema: {$ref: "#/components/schemas/CollectorEnrollment"}}}
        "400": {description: Invalid expiration}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Collector registry unavailable}
  /collectors/{id}/config:
    put:
      operationId: publishCollectorConfig
      x-required-permission: source:manage
      summary: Publish a new immutable Collector configuration version
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: '^col_[a-f0-9]{32}$'}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true, maxProperties: 128}
      responses:
        "201": {description: Configuration version created; inline secrets are rejected}
        "400": {description: Invalid configuration or inline credentials}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Collector not found}
        "503": {description: Collector registry unavailable}
  /collectors/{id}:
    delete:
      operationId: disableCollector
      x-required-permission: source:manage
      summary: Disable a Collector credential
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: '^col_[a-f0-9]{32}$'}}
      responses:
        "204": {description: Collector disabled}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Collector not found or already disabled}
        "503": {description: Collector registry unavailable}
  /collector/enroll:
    post:
      operationId: enrollCollector
      summary: Consume a one-time enrollment token and register this Collector
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enrollment_token, install_id, hostname, os, architecture, version]
              properties:
                enrollment_token: {type: string, maxLength: 128}
                install_id: {type: string, minLength: 1, maxLength: 128}
                hostname: {type: string, minLength: 1, maxLength: 255}
                os: {type: string, enum: [linux, windows]}
                architecture: {type: string, enum: [amd64, arm64]}
                version: {type: string, minLength: 1, maxLength: 64}
      responses:
        "201": {description: Collector ID and bearer credential returned once; response is no-store}
        "400": {description: Invalid registration}
        "401": {description: Invalid, expired, or consumed enrollment token}
        "503": {description: Collector registry unavailable}
  /collector/heartbeat:
    post:
      operationId: reportCollectorHeartbeat
      summary: Report Collector health and bounded source counters
      security:
        - collectorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CollectorHeartbeat"}
      responses:
        "204": {description: Heartbeat accepted}
        "400": {description: Invalid heartbeat}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {description: Collector registry unavailable}
  /collector/config:
    get:
      operationId: getCollectorConfig
      summary: Poll the latest immutable configuration using ETag
      security:
        - collectorBearer: []
      parameters:
        - {name: If-None-Match, in: header, required: false, schema: {type: string}}
      responses:
        "200": {description: Latest configuration version and ETag, no-store}
        "204": {description: No configuration has been published}
        "304": {description: Configuration is unchanged}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {description: Collector registry unavailable}
  /me:
    get:
      operationId: getCurrentPrincipal
      summary: Return the authenticated principal and effective permissions
      responses:
        "200":
          description: Current principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Principal"
        "401": {$ref: "#/components/responses/Unauthorized"}
  /overview:
    get:
      operationId: getOverview
      x-required-permission: anomaly:read
      summary: Return tenant-scoped investigation metrics
      responses:
        "200":
          description: Investigation overview
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Overview"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /events:
    get:
      operationId: listEvents
      x-required-permission: event:read
      summary: Search tenant-scoped normalized events in one supported domain
      description: Returns approved UIM fields only; original source payload is available through separately authorized evidence access.
      parameters:
        - {name: domain, in: query, required: true, schema: {type: string, enum: [authentication, session, iam, directory, network, dns, web, tls]}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Tenant and domain scoped event page
          content:
            application/json:
              schema:
                type: object
                required: [items, total, next_cursor]
                properties:
                  items: {type: array, items: {type: object}}
                  total: {type: integer}
                  next_cursor: {type: [string, "null"]}
        "400": {description: Invalid domain, time range, limit, or cursor}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Event store unavailable}
  /anomalies:
    get:
      operationId: listAnomalies
      x-required-permission: anomaly:read
      summary: List tenant-scoped anomalies
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - {name: severity, in: query, schema: {type: string, enum: [low, medium, high, critical]}}
        - {name: status, in: query, schema: {type: string, enum: [open, investigating, closed, false_positive]}}
        - {name: entity, in: query, schema: {type: string, maxLength: 128}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        "200":
          description: Anomaly page
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/AnomalySummary"}}
                  next_cursor: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /anomalies/{id}:
    get:
      operationId: getAnomaly
      x-required-permission: anomaly:read
      summary: Get a tenant-scoped anomaly
      parameters: [{name: id, in: path, required: true, schema: {type: string, maxLength: 256}}]
      responses:
        "200":
          description: Anomaly detail
          content: {application/json: {schema: {$ref: "#/components/schemas/AnomalyDetail"}}}
        "404": {description: Anomaly not found}
  /anomalies/{id}/evidence:
    get:
      operationId: listAnomalyEvidence
      x-required-permission: anomaly:read
      summary: Get the authentication events referenced by an anomaly
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, maxLength: 256}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
      responses:
        "200":
          description: Tenant-scoped event evidence
          content:
            application/json:
              schema:
                type: object
                required: [items, total, truncated]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/EvidenceEvent"}}
                  total: {type: integer}
                  truncated: {type: boolean}
        "404": {description: Anomaly not found}
  /cases:
    get:
      operationId: listCases
      x-required-permission: case:read
      summary: List tenant-scoped cases
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - {name: status, in: query, schema: {type: string, enum: [open, in_progress, closed]}}
        - {name: severity, in: query, schema: {type: string, enum: [low, medium, high, critical]}}
        - {name: q, in: query, schema: {type: string, maxLength: 200}}
      responses:
        "200": {description: Case page}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: createCase
      x-required-permission: case:write
      summary: Create a tenant-scoped case
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 8, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateCase"}
      responses:
        "201":
          description: Case created or the stored idempotent response
          content: {application/json: {schema: {$ref: "#/components/schemas/Case"}}}
        "409": {description: Idempotency key was reused with a different request}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /cases/{id}:
    get:
      operationId: getCase
      x-required-permission: case:read
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {"200": {description: Case detail}, "404": {description: Case not found}}
  /cases/{id}/actions:
    post:
      operationId: updateCase
      x-required-permission: case:write
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {"200": {description: Case updated}, "409": {description: Version conflict}}
...
  /members:
    get:
      operationId: listMembers
      x-required-permission: user:manage
      responses: {"200": {description: Tenant memberships}, "403": {$ref: "#/components/responses/Forbidden"}}
...
components:
  securitySchemes:
    sourceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    collectorBearer:
      type: http
      scheme: bearer
      bearerFormat: TUBA Collector credential
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/realms/tuba/.well-known/openid-configuration
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 2048}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
...
  schemas:
...
    SourceInstance:
      type: object
      required: [id, organization_id, namespace, vendor_name, vendor_product, vendor_dataset, source_epoch, source_context_id, release_id, enabled, rate_limit, created_at, updated_at]
      properties:
        id: {type: string}
        organization_id: {type: string}
        namespace: {type: string}
        vendor_name: {type: string}
        vendor_product: {type: string}
        vendor_dataset: {type: string}
        source_epoch: {type: string}
        source_context_id: {type: string, pattern: '^ctx_[a-f0-9]{32}$'}
        release_id: {type: string}
        enabled: {type: boolean}
        rate_limit: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...